+++ SOS/Utilities/Conversion.h
@@ -5,6 +5,6 @@
 //#include <stringapiset.h>
 
 std::string toString(double value);
-std::string toString(unsigned long value);
+std::string toString(unsigned int value);
 std::string wstrtostr(const std::wstring &wstr);
 std::wstring strtowstr(const std::string &str);
